Replace dashboard with feed and implement token refresh
