+++ frontend/src/main.jsx
@@ -4,7 +4,6 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import App from './App.jsx'
 import Login from './Login.jsx'
 import Registration from './Registration.jsx'
-import Dashboard from './Dashboard.jsx'
 import Feed from './Feed.jsx'
 import './index.css'
 
@@ -15,7 +14,6 @@ createRoot(document.getElementById('root')).render(
         <Route path="/" element={<App />} />
         <Route path="/login" element={<Login />} />
         <Route path="/registration" element={<Registration />} />
-        <Route path="/dashboard" element={<Dashboard />} />
         <Route path="/feed" element={<Feed />} />
       </Routes>
     </BrowserRouter>
